refactor: Main 컴포넌트 삭제 및 App에 통합
pages 폴더에서 page 컴포넌트 관리
