+++ src/components/pages/ProductCartPage/index.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 
-function ShoppingCartPage() {
+function ProductCartPage() {
   return <h2>👷🚧 아직 준비중~~^^ 🚧👷</h2>;
 }
 
-export default ShoppingCartPage;
+export default ProductCartPage;
