+++ src/components/pages/index.jsx
@@ -1,0 +1,6 @@
+import OrderListPage from "../pages/OrderListPage";
+import ProductListPage from "../pages/ProductListPage";
+import ProductDetailPage from "../pages/ProductDetailPage";
+import ProductCartPage from "../pages/ProductCartPage";
+
+export { OrderListPage, ProductListPage, ProductDetailPage, ProductCartPage };
