+++ src/components/styled.jsx
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+
+export const Main = styled.main`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  width: 100%;
+  margin: 140px auto 60px;
+
+  height: calc(100vh - 200px);
+`;
