+++ src/index.js
@@ -1,7 +1,8 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
-import App from "./App";
+import App from "./components/App";
+
 import GlobalStyle from "./GlobalStyle";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
